*** empty log message ***
Former-commit-id: 0227b27f85baa7b13d02beade07b5df982912f90

+++ PT/AMPS/src/general/cell.h
@@ -67,7 +67,7 @@ public:
 
   void GetCellCenter(float* x) {
     int i;
-    float center[3];
+    double center[3];
 
    GetCellCenter(center);
    for (i=0;i<DIM;i++) x[i]=center[i];
